+++ src/Model/Table/PostsTable.php
@@ -23,6 +23,7 @@
 namespace MeCms\Model\Table;
 
 use Cake\Cache\Cache;
+use Cake\I18n\Time;
 use Cake\ORM\Query;
 use Cake\ORM\RulesChecker;
 use MeCms\Model\Entity\Post;
@@ -38,9 +39,13 @@ class PostsTable extends AppTable {
 	 * @param \Cake\ORM\Entity $entity Entity object
 	 * @param \ArrayObject $options Options
 	 * @uses Cake\Cache\Cache::clear()
+	 * @uses setNextToBePublished()
 	 */
 	public function afterDelete(\Cake\Event\Event $event, \Cake\ORM\Entity $entity, \ArrayObject $options) {
-		Cache::clear(FALSE, 'posts');		
+		Cache::clear(FALSE, 'posts');	
+		
+		//Sets the next post to be published
+		$this->setNextToBePublished();	
 	}
 	
 	/**
@@ -49,9 +54,32 @@ class PostsTable extends AppTable {
 	 * @param \Cake\ORM\Entity $entity Entity object
 	 * @param \ArrayObject $options Options
 	 * @uses Cake\Cache\Cache::clear()
+	 * @uses setNextToBePublished()
 	 */
 	public function afterSave(\Cake\Event\Event $event, \Cake\ORM\Entity $entity, \ArrayObject $options) {
 		Cache::clear(FALSE, 'posts');
+		
+		//Sets the next post to be published
+		$this->setNextToBePublished();
+	}
+	
+	/**
+	 * Sets in cache the timestamp of the next post to be published.
+	 * This value can be used to check if the cache is valid
+	 * @uses Cake\I18n\Time::toUnixString()
+	 */
+	public function setNextToBePublished() {
+		$post = $this->find()
+			->select('created')
+			->where([
+				sprintf('%s.active', $this->alias())	=> TRUE,
+				sprintf('%s.created >', $this->alias()) => new Time()
+			])
+			->order([sprintf('%s.created', $this->alias()) => 'ASC'])
+			->first();
+		
+		if(!empty($post) && !empty($post->created))
+			Cache::write('nextToBePublished', $post->created->toUnixString(), 'posts');
 	}
 
     /**