+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.0 branch
 ### 2.0.1-beta
+* fixed bug, now the cache is flushed automatically, if there's a post-dated post to be published;
 * fixed a bug with the posts date.
 
 ### 2.0.0-alpha